+++ src/internet_identity/src/storage/account.rs
@@ -4,10 +4,7 @@ use internet_identity_interface::internet_identity::types::{
     AccountNumber, AnchorNumber, FrontendHostname, Timestamp,
 };
 use serde::Deserialize;
-use std::{
-    borrow::Cow,
-    hash::{DefaultHasher, Hash, Hasher},
-};
+use std::borrow::Cow;
 
 #[derive(
     Clone, Debug, Deserialize, CandidType, serde::Serialize, Eq, PartialEq, Ord, PartialOrd,
@@ -83,53 +80,4 @@ impl Storable for StorableAccount {
     const BOUND: Bound = Bound::Unbounded;
 }
 
-#[derive(Clone, Debug, Deserialize, serde::Serialize, Eq, PartialEq, Ord, PartialOrd)]
-pub struct Application {
-    pub origin: FrontendHostname,
-    pub total_accounts: u64,
-}
 
-impl Storable for Application {
-    fn to_bytes(&self) -> Cow<[u8]> {
-        Cow::Owned(serde_cbor::to_vec(&self).unwrap())
-    }
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        serde_cbor::from_slice(&bytes).unwrap()
-    }
-
-    const BOUND: Bound = Bound::Unbounded;
-}
-
-#[derive(Clone, Debug, Ord, PartialOrd, Eq, PartialEq)]
-pub struct OriginHash {
-    hash: [u8; 8],
-}
-
-impl Storable for OriginHash {
-    fn to_bytes(&self) -> Cow<[u8]> {
-        Cow::Owned(self.hash.to_vec())
-    }
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        Self {
-            hash: bytes.as_ref().try_into().unwrap(),
-        }
-    }
-
-    const BOUND: Bound = Bound::Bounded {
-        max_size: 8,
-        is_fixed_size: true,
-    };
-}
-
-impl OriginHash {
-    pub fn from_origin(origin: &FrontendHostname) -> Self {
-        let mut hasher = DefaultHasher::new();
-        origin.hash(&mut hasher);
-        let hash_u64 = hasher.finish();
-        Self {
-            hash: hash_u64.to_le_bytes(),
-        }
-    }
-}
