+++ src/internet_identity/src/storage/application.rs
@@ -1,0 +1,58 @@
+use ic_stable_structures::{storable::Bound, Storable};
+use internet_identity_interface::internet_identity::types::FrontendHostname;
+use serde::Deserialize;
+use std::{
+    borrow::Cow,
+    hash::{DefaultHasher, Hash, Hasher},
+};
+
+#[derive(Clone, Debug, Deserialize, serde::Serialize, Eq, PartialEq, Ord, PartialOrd)]
+pub struct Application {
+    pub origin: FrontendHostname,
+    pub stored_accounts: u64,
+}
+
+impl Storable for Application {
+    fn to_bytes(&self) -> Cow<[u8]> {
+        Cow::Owned(serde_cbor::to_vec(&self).unwrap())
+    }
+
+    fn from_bytes(bytes: Cow<[u8]>) -> Self {
+        serde_cbor::from_slice(&bytes).unwrap()
+    }
+
+    const BOUND: Bound = Bound::Unbounded;
+}
+
+#[derive(Clone, Debug, Ord, PartialOrd, Eq, PartialEq)]
+pub struct OriginHash {
+    hash: [u8; 8],
+}
+
+impl Storable for OriginHash {
+    fn to_bytes(&self) -> Cow<[u8]> {
+        Cow::Owned(self.hash.to_vec())
+    }
+
+    fn from_bytes(bytes: Cow<[u8]>) -> Self {
+        Self {
+            hash: bytes.as_ref().try_into().unwrap(),
+        }
+    }
+
+    const BOUND: Bound = Bound::Bounded {
+        max_size: 8,
+        is_fixed_size: true,
+    };
+}
+
+impl OriginHash {
+    pub fn from_origin(origin: &FrontendHostname) -> Self {
+        let mut hasher = DefaultHasher::new();
+        origin.hash(&mut hasher);
+        let hash_u64 = hasher.finish();
+        Self {
+            hash: hash_u64.to_le_bytes(),
+        }
+    }
+}
